// garland-house-frontend/src/fsd/entities/RelaysSwitches/model/types.ts
export type RelaysSwitchesType = {
  title: "Реле и выключатели";
  wireless_1: number;
  wireless_2: number;
  wireless_3: number;
  wireless_1_wifi: number;
  wireless_2_wifi: number;
  wireless_3_wifi: number;
  astroRelay: number;
  photoRelay: number;
  default_1: number;
  default_2: number;
  timeRelay: number;
  priceObj: {
    wireless_1: number;
    wireless_2: number;
    wireless_3: number;
    wireless_1_wifi: number;
    wireless_2_wifi: number;
    wireless_3_wifi: number;
    astroRelay: number;
    photoRelay: number;
    timeRelay: number;
  };
};